pixelnbrand-django/requirements.txt
Django>=4.2
mysqlclient>=2.1
Pillow>=10.0
python-decouple>=3.8
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
django-ckeditor